requirements.txt
pillow
pytesseract
tdqm
pyyaml
typer[all]